add dcos url and service url helpers
Prior to this commit, we didn't have a way to easily get the DCOS
cluster URL or a DCOS service's URL from Shakedown itself; we'd need to
use methods available in the 'dcos' library.

This commit introduces a couple helpers in 'shakedown.dcos',
specifically dcos_url() and dcos_service(url). These methods will return
strings with the cluster and service URLs, respectively, allowing users
to use these values in their tests. For example:

    from shakedown.dcos import dcos_service_url
    requests.post(dcos_service_url(jenkins ... ))