Remove 'key not found' error
By default, shakedown/paramiko will look for a valid key (default: `~/.ssh/id_rsa`), then fall back to those loaded in `ssh-agent`.  Previous to this patch, an error would be continually displayed if `~/.ssh/id_rsa` could not be found.  With this removal, the entire keychain will be examined and an error only displayed if no valid key is available for use.