Wait for data to be ready to read and add documentation
Add more python docstrings to HostSession handlers.

It is possible that when `recv_ready` is called there is no data ready
to be read, even though the connection is active and open. This can
cause the output from the command to never be populate. Instead,
first wait for data to be ready or the connection to be closed before
trying to read.