Replace print statements with logger calls. (#6641)
Summary:
This patch replaces most `print` calls with a `logger.info` call. Only
the spinner and a command flush call uses print. The spinner will be
removed with MARATHON-8387.

This also removed Shakedown's unit tests as they are not run and
outdated. We test Shakedown implicitly in our system tests.

JIRA issues: MARATHON-8416